hw-3 refactor routes, add new entity

+++ src/crud/db/user.dal.ts
@@ -3,6 +3,7 @@ import UserModel from '../models/user.model';
 import { UserInput, UserOutput, UsersFilters } from '../types';
 
 const NO_USER_MSG = 'No such user';
+const EMPTY_STR_MSG = 'Empty string';
 
 export const create = async (payload: UserInput): Promise<UserOutput> => {
   const user = await UserModel.create(payload);
@@ -27,6 +28,10 @@ export const update = async (id: string, payload: UserInput): Promise<UserOutput
 };
 
 export const deleteById = async (id: string): Promise<boolean> => {
+  const user = await UserModel.findByPk(id);
+  if (user == null) {
+    throw new Error(NO_USER_MSG);
+  }
   const deletedUserCount = await UserModel.destroy({
     where: { id },
   });
@@ -34,11 +39,14 @@ export const deleteById = async (id: string): Promise<boolean> => {
 };
 
 export const getAll = async (filters?: UsersFilters): Promise<UserOutput[]> => {
-  const isLoginSubstring = filters?.loginSubstring !== undefined;
+  const isLoginSubstring = filters?.loginSubstring !== '';
   const isLimit = filters?.limit !== undefined;
+  if (!isLoginSubstring) {
+    throw new Error(EMPTY_STR_MSG);
+  }
   const users = await UserModel.findAll({
     where: {
-      ...(isLoginSubstring && { login: { [Op.substring]: filters.loginSubstring } }),
+      ...(isLoginSubstring && { login: { [Op.substring]: filters?.loginSubstring } }),
     },
     ...(isLimit && { limit: filters.limit }),
     order: [['login', 'ASC']],
